Correcciones del backend

+++ backend/src/controllers/mantencion.controller.js
@@ -49,7 +49,7 @@ const getMantencionById = async (req, res) => {
         message: 'Mantencion no existe'
       });
     }
-  
+
     return res.json(mantencion);
   } catch (error) {
     return res.status(500).json({
@@ -60,27 +60,38 @@ const getMantencionById = async (req, res) => {
 
 const updateMantencionById = async (req, res) => {
   try {
-    const mantencionActualizada = await Mantencion.findByIdAndUpdate(req.params.id, {
-      descripcion: req.body.descripcion,
-      instalacion: req.body.instalacion,
-      fecha_inicio: req.body.fecha_inicio,
-      fecha_termino: req.body.fecha_termino,
-      imagen_antes: req.files.imagen_antes && req.files.imagen_antes[0].path,
-      imagen_despues: req.files.imagen_despues && req.files.imagen_despues[0].path,
-      empresa: req.body.empresa,
-      costo: req.body.costo,
-      rutinaria: req.body.rutinaria,
-      estado: req.body.estado
-    }, {
-      new: true
-    });
+    let mantencion = await Mantencion.findById(req.params.id);
 
-    if (!mantencionActualizada) {
+    if (!mantencion) {
       return res.status(404).json({
         message: 'Mantencion no existe'
       });
     }
-  
+
+    mantencion.descripcion = req.body.descripcion;
+    mantencion.instalacion = req.body.instalacion;
+    mantencion.fecha_inicio = req.body.fecha_inicio;
+    mantencion.fecha_termino = req.body.fecha_termino;
+    mantencion.empresa = req.body.empresa;
+    mantencion.costo = req.body.costo;
+    mantencion.rutinaria = req.body.rutinaria;
+    mantencion.estado = req.body.estado;
+
+    if (req.body.imagen_antes === 'null') {
+      await fs.unlink(path.resolve(mantencion.imagen_antes));
+      mantencion.imagen_antes = undefined;
+    } else if (req.files.imagen_antes) {
+      mantencion.imagen_antes = req.files.imagen_antes[0].path;
+    }
+    if (req.body.imagen_despues === 'null') {
+      await fs.unlink(path.resolve(mantencion.imagen_despues));
+      mantencion.imagen_despues = undefined;
+    } else if (req.files.imagen_despues) {
+      mantencion.imagen_despues = req.files.imagen_despues[0].path;
+    }
+
+    const mantencionActualizada = await mantencion.save();
+
     return res.json({
       message: 'Mantencion actualizado correctamente',
       mantencion: mantencionActualizada
@@ -94,17 +105,20 @@ const updateMantencionById = async (req, res) => {
 
 const deleteMantencionById = async (req, res) => {
   try {
-      const mantencionEliminada = await Mantencion.findByIdAndDelete(req.params.id);
+    const mantencionEliminada = await Mantencion.findByIdAndDelete(req.params.id);
 
     if (!mantencionEliminada) {
       return res.status(404).json({
         message: 'Mantencion no existe'
       });
     }
+    if (mantencionEliminada.imagen_antes) {
+      await fs.unlink(path.resolve(mantencionEliminada.imagen_antes));
+    }
+    if (mantencionEliminada.imagen_despues) {
+      await fs.unlink(path.resolve(mantencionEliminada.imagen_despues));
+    }
 
-    await fs.unlink(path.resolve(mantencionEliminada.imagen_antes));
-    await fs.unlink(path.resolve(mantencionEliminada.imagen_despues));
-  
     return res.json({
       message: 'Mantencion eliminada correctamente'
     });
@@ -116,9 +130,9 @@ const deleteMantencionById = async (req, res) => {
 };
 
 module.exports = {
-    createMantencion,
-    getMantenciones,
-    getMantencionById,
-    updateMantencionById,
-    deleteMantencionById
+  createMantencion,
+  getMantenciones,
+  getMantencionById,
+  updateMantencionById,
+  deleteMantencionById
 }
